requirements.txt
pillow
einops
spaces
torch==2.3.1
timm
transformers
supervision==0.22.0
gradio